Return risk assessment on form submit.
Classification is based on the percentage difference between the input
risk and the maximum possible risk for a given request. Fields with
multi-valued inputs are handled by diffing the maximum possible risk for
that field with the number of input values in the current request with
the summed risks of the inputs.